Move StatsWindow class to it's own file.
